src/requirements.txt
openai==2.11.0
requests>=2.30.0
python-dotenv>=1.0.0
pg8000>=1.29.0
pydantic<2.0.0